I did CMake and built a run file for the whole project and corrected errors that were created as a result.
