+++ Security/CMakeLists.txt
@@ -1,0 +1,14 @@
+cmake_minimum_required(VERSION 3.5)
+project(team_Security)
+set(CMAKE_CXX_STANDARD 17)
+set(CMAKE_CXX_STANDARD_REQUIRED ON)
+file(GLOB SOURCE_FILES "Security/*.cpp")
+file(GLOB HEADER_FILES "Security/*.h")
+add_executable(team_Security ${SOURCE_FILES} ${HEADER_FILES})
+find_package(Boost REQUIRED)
+if(Boost_FOUND)
+ include_directories(${Boost_INCLUDE_DIRS})
+ target_link_libraries(team_Security ${Boost_LIBRARIES})
+else()
+    message(FATAL_ERROR "Boost not found!")
+endif()
