+++ Security/Security/KeyPair.h
@@ -10,17 +10,19 @@ class KeyPair {
 
 private:
     cpp_int privateKey;
-    cpp_int publicKey;
+    cpp_int publicKey_RSA;
     Point publicKey;
+
 public:
     KeyPair(const EllipticCurve& curve);
-    cpp_int getPrivateKey() const;
-    Point getPublicKey() const;
-    cpp_int generate_random_cpp_int(int bit_length);
-    cpp_int generate_prime(int bit_length);
-    cpp_int gcd(cpp_int a, cpp_int b);
-    void set_private_and_public_keys();
 
-}; 
+    cpp_int getPrivateKey() const;
+    cpp_int getPublicKey_RSA() const;
+    Point getPublicKey() const;
+
+    void setPrivateKey(cpp_int privateKey);
+    void setPublicKey_RSA(cpp_int publicKey_RSA);
+    void setPublicKey(Point publicKey);
+};
 
 #endif // KEY_PAIR_H
